requirements.txt
fastapi
requests
google-generativeai